+++ lib/src/internal/layout/drop_item.dart
@@ -180,7 +180,7 @@ class DropItem extends LayoutModifier {
           children.add(newChild);
         }
       });
-      if (children.length == 0) {
+      if (children.isEmpty) {
         return null;
       } else if (children.length == 1) {
         return children.first;